+++ src/api/objects/account.ts
@@ -9,7 +9,7 @@ import MastoApi from "../api";
 import MastodonServer, { InstanceResponse } from '../mastodon_server';
 import { config } from "../../config";
 import { DEFAULT_FONT_SIZE, bracketed, extractDomain, replaceEmojiShortcodesWithImgTags } from "../../helpers/string_helpers";
-import { keyByProperty } from "../../helpers/collection_helpers";
+import { groupBy, keyByProperty } from "../../helpers/collection_helpers";
 import { Logger } from "../../helpers/logger";
 import { type AccountLike, type AccountNames, type BooleanFilterOption, type StringNumberDict } from "../../types";